requirements.txt
fastapi==0.127.0
uvicorn[standard]==0.40.0

sqlmodel==0.0.28
psycopg[binary]==3.3.2

python-dotenv==1.2.1
pydantic-settings==2.10.1
email-validator==2.2.0

passlib[bcrypt]==1.7.4
bcrypt==4.0.1

PyJWT==2.10.1

slowapi==0.1.9
limits==5.6.0

pytest==8.4.2
httpx==0.28.1
